+++ typescript/vrotsc/src/compiler/transformer/helpers/source.ts
@@ -20,24 +20,6 @@ export function addHeaderComment(statements: ts.Statement[]): void {
 	}
 }
 
-/**
-* Maps a TypeScript type to a VRO type.
-* @TODO: What is it used for?
-*/
-export function getVroType(typeNode: ts.TypeNode): string {
-	switch (typeNode.kind) {
-		case ts.SyntaxKind.StringKeyword:
-			return "string";
-		case ts.SyntaxKind.NumberKeyword:
-			return "number";
-		case ts.SyntaxKind.BooleanKeyword:
-			return "boolean";
-		case ts.SyntaxKind.ArrayType:
-			return "Array/" + getVroType((<ts.ArrayTypeNode>typeNode).elementType);
-		default:
-			return "Any";
-	}
-}
 
 /**
 * Prints a source file to a string.
