--- typescript/vrotsc/tmp/compiler/elementInfo.ts
@@ -1,28 +1,0 @@
-/*
- * #%L
- * vrotsc
- * %%
- * Copyright (C) 2023 VMware
- * %%
- * Build Tools for VMware Aria
- * Copyright 2023 VMware, Inc.
- * 
- * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.  
- * 
- * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
- * #L%
- */
-namespace vrotsc {
-    export function printElementInfo(properties: Record<string, string>): string {
-        const stringBuilder = createStringBuilder();
-        stringBuilder.append(`<?xml version="1.0" encoding="utf-8" standalone="no"?>`).appendLine();
-        stringBuilder.append(`<!DOCTYPE properties SYSTEM "http://java.sun.com/dtd/properties.dtd">`).appendLine();
-        stringBuilder.append(`<properties>`).appendLine();
-        stringBuilder.append(`<comment>Generated by vrotsc</comment>`).appendLine();
-        Object.keys(properties).forEach(key => {
-            stringBuilder.append(`<entry key="${key}">${properties[key]}</entry>`).appendLine();
-        });
-        stringBuilder.append(`</properties>`).appendLine();
-        return stringBuilder.toString();
-    }
-}
